fix(tickets): preserve document boundaries and test isolation
