+++ packages/tickets/src/components/ticket/__tests__/TicketDetails.remoteUpdates.test.tsx
@@ -41,11 +41,15 @@ let liveTicketContext = {
   reconnectVersion: 0,
 };
 
-vi.mock('next/navigation', () => ({
-  useRouter: () => ({ push: routerPushMock, refresh: vi.fn() }),
-  useSearchParams: () => new URLSearchParams(),
-  usePathname: () => '/msp/tickets/ticket-1',
-}));
+vi.mock('next/navigation', () => {
+  // A fresh router on each render would reset the remote-update debounce.
+  const router = { push: routerPushMock, refresh: vi.fn() };
+  return {
+    useRouter: () => router,
+    useSearchParams: () => new URLSearchParams(),
+    usePathname: () => '/msp/tickets/ticket-1',
+  };
+});
 
 vi.mock('next-auth/react', () => ({
   useSession: () => ({ data: { user: { id: 'user-1' } } }),
@@ -162,29 +166,34 @@ vi.mock('@alga-psa/ui/ui-reflection/ReflectionContainer', () => ({
   ReflectionContainer: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
 }));
 
-vi.mock('@alga-psa/ui/lib/i18n/client', () => ({
-  // Components under test format dates through useFormatters; the real hook
-  // reads the locale off the provider this test does not mount.
-  useFormatters: () => ({
-    locale: 'en',
-    formatDate: (date: Date | string, options?: Intl.DateTimeFormatOptions) =>
-      new Intl.DateTimeFormat('en', options).format(typeof date === 'string' ? new Date(date) : date),
-    formatNumber: (value: number, options?: Intl.NumberFormatOptions) =>
-      new Intl.NumberFormat('en', options).format(value),
-    formatCurrency: (value: number, currency: string, options?: Intl.NumberFormatOptions) =>
-      new Intl.NumberFormat('en', { style: 'currency', currency, ...options }).format(value),
-    formatRelativeTime: (date: Date | string) => String(date),
-  }),
-  useTranslation: () => ({
-    t: (_key: string, fallback?: string | Record<string, unknown>) => {
-      // Mirror i18next's t(key, options) form where options carries defaultValue.
-      if (fallback && typeof fallback === 'object') {
-        fallback = typeof fallback.defaultValue === 'string' ? fallback.defaultValue : undefined;
-      }
-      return typeof fallback === 'string' ? fallback : _key;
-    },
-  }),
-}));
+vi.mock('@alga-psa/ui/lib/i18n/client', () => {
+  // Keep t stable like the provider; replacing it on every render requeues the
+  // remote-update debounce while asynchronous mount effects are settling.
+  const translate = (_key: string, fallback?: string | Record<string, unknown>) => {
+    // Mirror i18next's t(key, options) form where options carries defaultValue.
+    if (fallback && typeof fallback === 'object') {
+      fallback = typeof fallback.defaultValue === 'string' ? fallback.defaultValue : undefined;
+    }
+    return typeof fallback === 'string' ? fallback : _key;
+  };
+  return {
+    // Components under test format dates through useFormatters; the real hook
+    // reads the locale off the provider this test does not mount.
+    useFormatters: () => ({
+      locale: 'en',
+      formatDate: (date: Date | string, options?: Intl.DateTimeFormatOptions) =>
+        new Intl.DateTimeFormat('en', options).format(typeof date === 'string' ? new Date(date) : date),
+      formatNumber: (value: number, options?: Intl.NumberFormatOptions) =>
+        new Intl.NumberFormat('en', options).format(value),
+      formatCurrency: (value: number, currency: string, options?: Intl.NumberFormatOptions) =>
+        new Intl.NumberFormat('en', { style: 'currency', currency, ...options }).format(value),
+      formatRelativeTime: (date: Date | string) => String(date),
+    }),
+    useTranslation: () => ({
+      t: translate,
+    }),
+  };
+});
 
 vi.mock('@alga-psa/tags/context', () => ({
   useTags: () => ({ tags: [] }),
